% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.classify_sample}
\alias{.classify_sample}
\title{Helper function to classify a single sample}
\usage{
.classify_sample(sample, potential_distrs)
}
\arguments{
\item{sample}{sample observations}

\item{potential_distrs}{The types of distributions to fit}
}
\value{
character
}
\description{
Helper function to classify a single sample
}
\keyword{internal}
