% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenorm.R
\name{.get_normalisation}
\alias{.get_normalisation}
\title{Get values goal cuts}
\usage{
.get_normalisation(x, brks, polarity)
}
\arguments{
\item{x}{A vector of values}

\item{brks}{The breaks to normalise to}

\item{polarity}{Which direction should the normalisation occur.}
}
\description{
Get values goal cuts
}
\keyword{internal}
