% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.recommend}
\alias{.recommend}
\title{Helper function to check for recommendations}
\usage{
.recommend(x, distr, outliers, classint_pref, nclasses, potential_distrs)
}
\arguments{
\item{x}{The observations}

\item{distr}{The likely distribution}

\item{outliers}{Does the data have IQR outliers}

\item{classint_pref}{The preferred classInt style}

\item{nclasses}{The number of desired classes for classInt}

\item{potential_distrs}{The types of distributions to fit}
}
\value{
list with the following description:
\itemize{
\item norm: character() the recommended normalisation technique
\item breaks: numeric The recommended breaks
\item mdl: the \code{univariateML} model
}
}
\description{
Helper function to check for recommendations
}
\keyword{internal}
