% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenorm.R
\name{piecenorm}
\alias{piecenorm}
\title{Get values bins}
\usage{
piecenorm(x, brks, polarity = 1)
}
\arguments{
\item{x}{A vector of values}

\item{method}{The scaling method to use}
}
\description{
Get values bins
}
