% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenorm.R
\name{piecewise_normalisation}
\alias{piecewise_normalisation}
\title{Get values bins}
\usage{
piecewise_normalisation(x, brks, polarity = 1)
}
\arguments{
\item{x}{A vector of values}

\item{method}{The scaling method to use}
}
\description{
Get values bins
}
