% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenorm.R
\name{.calculate_rescaled}
\alias{.calculate_rescaled}
\title{Calculate the rescaling}
\usage{
.calculate_rescaled(tmp, polarity)
}
\arguments{
\item{tmp}{A preprocessed data frame}

\item{polarity}{Which direction should the normalisation occur.}
}
\description{
Calculate the rescaling
}
\keyword{internal}
