% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecenorm.R
\name{.get_normalisation}
\alias{.get_normalisation}
\title{Get values goal cuts}
\usage{
.get_normalisation(obs, breaks)
}
\arguments{
\item{obs}{A vector of observations}

\item{breaks}{The breaks to normalise to}
}
\description{
Get values goal cuts
}
\keyword{internal}
